Fixed integer casts in dll unit tests.

+++ test/doubly_linked_list/test_doubly_linked_list_insert_beginning.cpp
@@ -60,8 +60,8 @@ TEST_F(dll_insert_beginning_test, basic_test)
     // the number of elements should be 2, with the first element being the
     // second data item and the last element being the first data item
     EXPECT_EQ(dll.elements, 2UL);
-    EXPECT_EQ(*(long*)(dll.first->data), data2);
-    EXPECT_EQ(*(long*)(dll.last->data), data);
+    EXPECT_EQ(*(int*)(dll.first->data), data2);
+    EXPECT_EQ(*(int*)(dll.last->data), data);
 
     // test the links on each element
     EXPECT_EQ(dll.first->prev, nullptr);
